refactor: extract helpers from index.ts into module structure (Phase 1)
Creates ServerContext interface and extracts ~845 lines of helper functions
into dedicated modules:

- src/server-context.ts: ServerContext interface with shared singletons
- src/helpers/project.ts: resolveProject, ensureBranchSynced, etc.
- src/helpers/git-commit.ts: formatCommitBody, shouldBlockProtectedBranchCommit, etc.
- src/helpers/embed.ts: embedTextForNote, embedMissingNotes, etc.
- src/helpers/persistence.ts: buildMutationRetryContract, pushAfterMutation, etc.
- src/helpers/vault.ts: storageLabel, collectVisibleNotes, moveNoteBetweenVaults, etc.
- src/helpers/index.ts: slugify, formatNote, formatTemporalHistory, etc.

index.ts retains wrapper functions that close over ctx so tool handlers remain
unchanged. No functional changes — pure structural refactoring.

All 845 tests pass, tsc --noEmit clean.
